benchupload: Collect xunit files in travis-ci projects first
Begin with a structure for CI systems, searching xunit files and
then upload them to the server. data was picked over custom headers
to avoid having a http proxy remove these header values.

In the future we will store any sort of key/value/unit data and
have private tokens for the upload.